Refactored to match scaffolding branch
